KUDU-3297 fix RPC negotiations with cyrus-sasl-gssapi-2.1.27-5 and newer
It turns out that setting SASL security properties such as the minimum
Security Strength Factor (SSF) once sasl_client_start() has already
been called isn't working as expected anymore once patch [1] is applied
for the cyrus-sasl-gssapi plugin.

This patch addresses the issue, moving the call to
sasl_setprop(..., SASL_SEC_PROPS, ...) prior to the corresponding call
to sasl_client_start() in the client-side negotiation logic for C++ Kudu
components.

Prior to this patch, GSSAPI-involved scenarios of the negotiation-test
and security-itest would fail when running against the GSSAPI plugin
with patch [1] applied.

With this patch, all scenarios in the negotiation-test and the
security-itest pass.

I didn't add any extra test scenarios since the already existing test
coverage was enough to spot the issue, as it can be seen from above.

[1] https://github.com/cyrusimap/cyrus-sasl/pull/603

Change-Id: Ia655356798c753d5a223933cc09a0731018e10af
Reviewed-by: Grant Henke <ghenke@cloudera.com>
Reviewed-on: http://gerrit.cloudera.org:8080/17620
Reviewed-by: Joe McDonnell <joemcdonnell@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>