IMPALA-10765: Allow firing INSERT events with empty results
When INSERT or the INSERT part of a CTAS statement produces empty
results, catalogd will hit a Precondition check in createInsertEvents()
if the statement is not specified with OVERWRITE. This patch removes the
checks to relax the restriction. This is consistent with Hive since Hive
also fires insert events on empty insert results.

Previously when event processor is disabled, we won't generate INSERT
events. This is inconsistent with other DDL events which will always be
generated by Hive. This patch removes the check on event processor's
status. So firing INSERT events is independent to the event processor.
This fixes IMPALA-10744.

Tests
 - Add e2e test

Change-Id: Id708f38302b37d7e09435e7b8b0c9fc071659606
Reviewed-on: http://gerrit.cloudera.org:8080/17629
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>