IMPALA-10763: Min/max filters should be enabled on Z-order sorted columns
This patch enables min/max filtering on any Z-order sort-by columns
by default.

Since the column stats for a row group or a page is computed from the
column values stored in the row group or the page, the current
infrastructure for min/max filtering works for the Z-order out of box.
The fact that these column values are ordered by Z-order is
orthogonal to the work of min/max filtering.

By default, the new feature is enabled. Set the existing control knob
minmax_filter_sorted_columns to false to turn it off.

Testing
  1. Added new z-order related sort column tests in
     overlap_min_max_filters_on_sorted_columns.test;
  2. Ran core-test.

Change-Id: I2a528ffbd0e333721ef38b4be7d4ddcdbf188adf
Reviewed-on: http://gerrit.cloudera.org:8080/17635
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>