IMPALA-10766: Better selectivity for =,not distinct
For = :
If the right side is null, then selectivity is 0.
If the left side is null, null should be excluded when calculating
selectivity.

For is not distinct from :
If the right side is null, non null should be excluded when calculating
selectivity, and only null should be included.
If the left side is null and the right side is not null, null should be
excluded when calculating selectivity, including part of non-null.

Testing :
Change the UT, modify the selectivity calculation error, add two new
cases column != null and column = null

Change-Id: Ib8ec62f2355a7036125cc0d261b790644b9f4b60
Reviewed-on: http://gerrit.cloudera.org:8080/17637
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Qifan Chen <qchen@cloudera.com>