IMPALA-10774: Collect the output when starting Hive Metastore or HiveServer2
Before this patch, when we started the Hive Metastore or the HiveServer2
via run-hive-server.sh, hive-metastore.out and hive-server2.out would be
generated, respectively. The output of executing the hive command was
redirected to the corresponding files mentioned above. However, since
">" was used to redirect the output, the previously generated files
would be overwritten every time during the restart and thus some
important error message may be lost.

To collect more information regarding the restart, this patch appends
the newly produced output to hive-metastore.out and hive-server2.out
during the restart, which makes it easier for developers to troubleshoot
issues related to the restart.

Change-Id: I2efdbcf886e2d32ccf8c7eef038360884e44f216
Reviewed-on: http://gerrit.cloudera.org:8080/17642
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>