IMPALA-10490: Fix illegalStateException in drop stats
If a table was created and incremental stats were computed on the table
while events processing is turned off, then after the events processing
is turned on a drop stats or truncate table command on such a table
fails with IllegalStateException. This happens because the
catalog service identifiers are not found in the table properties
while partition for the table are being altered. This patch adds
the catalog service identifiers for drop stats, truncate and comment on
statement code paths to fix the error.

Testing:
1. Added more statements to the test_self_events test to cover the
newly added logic.
2. Manually tested the following scenario which was previously failing
  a. starting catalogd without events processing.
  b. create partitioned table and compute incremental stats on it.
  c. restart catalogd with events processing turned on.
  d. issue drop stats command.

Change-Id: Iaa0b4043879370c52049d22acb49c847b0be1c68
Reviewed-on: http://gerrit.cloudera.org:8080/17659
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>