IMPALA-10788: [DOCS] Introduce more flags for Statestore Scalability
statestore_subscriber_timeout_secs is also an important flag for
statestore scalability. Especially when the heartbeat frequency and
timeout are bumpped up, this flag should also be bumpped as well.
To introduce this flag, we should also introduce the heartbeat tcp
timeout flag and the max missed heartbeat flag.

Tests:
 - Built locally, verify the html and pdf artifacts are expected.

Change-Id: Ia4b331693c5c0945f4cec8fd81ed9ec688563333
Reviewed-on: http://gerrit.cloudera.org:8080/17675
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Shajini Thayasingh <sthayasingh@cloudera.com>
Reviewed-by: Tamas Mate <tmate@cloudera.com>