IMPALA-10799: Analysis slowdown with inline views and thousands of column
If there are thousands of columns in the inlineview, it‘s very slow in
analysis. Most of the cost is in the get() calls used to find
expressions in the local substitution map when checking if the column
is ambiguous.

The fix is to
1.Use LinkedHashMap to search and check if we have already seen the alias.
2.Do the check of checkComposedFrom() when the log level is TRACE since
the codes have been mature for a while.

Testing:
Performance testing with a query with 10000 expressions of the
following form:
  with a as (select c1 c1, c1 c2, c1 c3, ... from t)
  select c1, c2, c3, ... from a;
repro query analysis went from 7.5 sec to less than 1 sec.

Change-Id: I43da47dddfdb3db6d0e2073ae974a0a4d1b3ad7c
Reviewed-on: http://gerrit.cloudera.org:8080/17688
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>