IMPALA-10502: Fetch events in batches (Addendum)
The earlier change for IMPALA-10502 passes in a batch size
of -1 to fetch all the events from a given event id during a
DDL execution. While this works when HMS backing database is
postgres, it doesn't work well when the HMS backend
is a MySQL database due to HIVE-20226. This change works around the hive
bug to fetch the events in batches of 1000 instead of fetching all the events
in one RPC during the DDL execution.

Testing:
1. Added a unit test for the new changes introduced.
2. Ran the previously failing tests on MySQL HMS backend.

Change-Id: I34bb8984aeb91b37439f77722746f638d8774478
Reviewed-on: http://gerrit.cloudera.org:8080/17698
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>
Tested-by: Zoltan Borok-Nagy <boroknagyz@cloudera.com>