IMPALA-10821 Fix TestTPCHJoinQueries.test_outer_joins failed
A new added testcase in TestTPCHJoinQueries.test_outer_joins
can't pass the s3 build because the plan generated in s3 build
is not the same as the default hdfs build for the scan node. In
s3 build, it is "SCAN S3", while in hdfs build, it is "SCAN HDFS".

The patch changed the testcase to use $FILESYSTEM_NAME, which
will be changed according to the file system the testcase is
using.

Tests:
Reran and passed the failed testcase in s3 build.

Change-Id: I7e068d9da03517f8316e7a2505ce1466523d5917
Reviewed-on: http://gerrit.cloudera.org:8080/17716
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>