IMPALA-10858: HMS request for acid fds fixed.
This commit addresses a problem for full acid tables. When the
file descriptors are requested on an HMS call, the insert and
delete file descriptors need to be gathered up and sent back
to the caller.

Also changed the visibility of some methods in SingleNodePlanner
and an object within HdfsPartition to allow extensibility by
a third party tool.

Change-Id: I055960bf0471d62bc37dda3403f58f86f9b5f596
Reviewed-on: http://gerrit.cloudera.org:8080/17760
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>