IMPALA-8795: (Addendum) Fix test_hms_integration flakiness
The previous commit for IMPALA-8795 modifies the
test_sanity test to account for event processing is enabled.
However, it missed removing a assert statement before
event processing check. This commit modifies the test
to remove the unnecessary assert which fails intermittently.

Testing:
Looped the test 25 times and make sure it works with
default configs.

Change-Id: I647e6a37dbfecd6a4500c8178ecc07e605cbcfba
Reviewed-on: http://gerrit.cloudera.org:8080/17767
Reviewed-by: <kishen@cloudera.com>
Reviewed-by: Andrew Sherman <asherman@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>