IMPALA-9057: Disable test_transactional_insert_events
The test test_transactional_insert_events fails
intermittently. The root cause of the flakiness is
documented in the JIRA. Disabling this test until we
have the fix for the underlying cause of the problem.

Testing:
Confirmed that the test is skipped and marked as XFAIL
when test_events_processing is run.

Change-Id: I72de2c98d90036337a5fe12e652991b6f262d570
Reviewed-on: http://gerrit.cloudera.org:8080/17788
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>