IMPALA-10822: Allow multiple group returns for LDAP Search
This commit updates the LdapSearch method to allow multiple results by
refactoring the condition that checks the number of results. The
condition was removed from the method and it is now the caller's
responsibility to decide if the returned number of results are correct.

The change was necessary, because the LdapSearch method was written
too generally and used for both user and group searches. While it is
reasonable to fail authentication requests where the LDAP Search
returns more than one user this should not be the case for group
searches, as a user can be part of multiple groups matching the search
criteria.

Testing:
 - Updated the unit test to cover this use-case

Change-Id: I27eaf717985207e36d4fa47a2af757d06544474e
Reviewed-on: http://gerrit.cloudera.org:8080/17794
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>