IMPALA-10883: Do not override existing counters with empty profile
Some profile information was missing when gen_experimental_profile flag
is enabled. This is because profile aggregation in the coordinator does
not anticipate a partial update from a backend. From a backend
perspective, if a fragment instance has sent its final report, that
instance will not participate in the subsequent profile report.
Therefore, the aggregated counters that belong to the finished instance
will be empty. This patch adds empty checks in the aggregation of
input_profile_names_, TAggTimeSeriesCounter, and TAggEventSequence to
prevent the existing value from being overridden by an empty profile.

Testing:
- Add BE test CountersTest.PartialUpdate

Change-Id: I9bb179bf739ffaa4e5ec8dc911480ac835ae387f
Reviewed-on: http://gerrit.cloudera.org:8080/17819
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>