IMPALA-10885: Deflake test_get_table_req_without_fallback
The test originally was written when events processing
was not turned on by default. However, after
IMPALA-8795 the events processing is turned on by
default and this test fails intermittently.

The error was reproduced by adding a simple
sleep statement in the test just before issuing a
get_table API call which was expected to fail.

Testing:
1. Looped the test for 25 times with the change and with
the sleep statement which reproduced the failure.

Change-Id: I684ec07cc23617d64355df25420c45b0cbedd5a3
Reviewed-on: http://gerrit.cloudera.org:8080/17817
Reviewed-by: Vihang Karajgaonkar <vihang@cloudera.com>
Reviewed-by: Quanlong Huang <huangquanlong@gmail.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>