IMPALA-10905: test_iceberg.py test_time_travel fails in exhaustive builds
test_iceberg.py::test_time_travel sets the TIMEZONE query option
during execution. In the end it sets the TIMEZONE to 'Asia/Tokyo'.

The problem is that it sets the TIMEZONE of the client of the
ImpalaTestSuite. This means that a subsequent test will use
timezone 'Asia/Tokyo' instead of the local one.

This patch fixes this issue by creating a local Impala client for
the test.

Change-Id: I5c172df865bc35dd1792146f4045fbc16e04d116
Reviewed-on: http://gerrit.cloudera.org:8080/17835
Reviewed-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>
Tested-by: Impala Public Jenkins <impala-public-jenkins@cloudera.com>